+++ src/POS.Domain/Entities/Customer.cs
@@ -10,7 +10,7 @@ namespace POS.Domain.Entities
         public string LastName { get; set; } = string.Empty;
         public string ImageProfile { get; set; } = string.Empty;
         public string PhoneNumber { get; set; } = string.Empty;
-        public int TotalPoint { get; set; } = 0;
+        public decimal TotalPoint { get; set; } = 0;
         public bool Status { get; set; } = true;
         public virtual Person? Person { get; set; }
     }
